requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
coverage==4.5.1
defusedxml==0.5.0         # via python3-openid
django-allauth==0.36.0
django-cors-middleware==1.3.1
django-filter==1.1.0
django-nose==1.4.5
django-rest-auth==0.9.3
django==2.0.5
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
faker==0.8.15
flake8==3.5.0
idna==2.6                 # via requests
mccabe==0.6.1             # via flake8
nose==1.3.7               # via django-nose
oauthlib==2.1.0           # via requests-oauthlib
psycopg2==2.7.4
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
pyjwt==1.6.4              # via djangorestframework-jwt
python-dateutil==2.7.3    # via faker
python3-openid==3.1.0     # via django-allauth
pytz==2018.4              # via django
raven==6.8.0
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4          # via django-allauth, requests-oauthlib
six==1.11.0               # via django-rest-auth, faker, python-dateutil
text-unidecode==1.2       # via faker
urllib3==1.22             # via requests
